- update config drive test to not use --file since it is deprecated and no longer available in nova client in stein - update keywords/tests to get segmentation id from neutron net-show since behavior has changed - remove unnecessary teardown step in auto installer to reconnect via telnet > this can take up to 20 minutes and fails due to port got reset

+++ CGCSAuto/testcases/functional/networking/test_multiple_ports.py
@@ -141,6 +141,9 @@ class TestMutiPortsBasic:
             else:
                 LOG.tc_step("Perform following action(s) on vm {}: {}".format(vm_under_test, vm_actions))
                 for action in vm_actions:
+                    if 'migrate' in action and system_helper.is_simplex():
+                        continue
+
                     kwargs = {}
                     if action == 'hard_reboot':
                         action = 'reboot'
